+++ src/jpgimage.cpp
@@ -230,7 +230,7 @@ namespace Exiv2 {
             us2Data(tmpBuf + 4, iptc_, bigEndian);
             tmpBuf[6] = 0;
             tmpBuf[7] = 0;
-            ul2Data(tmpBuf + 8, (uint32_t)rawIptc.size_, bigEndian);
+            ul2Data(tmpBuf + 8, rawIptc.size_, bigEndian);
             append(psBlob, tmpBuf, 12);
             append(psBlob, rawIptc.pData_, rawIptc.size_);
             // Data is padded to be even (but not included in size)
